Load track with Once to not WaitGroup.Done twice
go-libspotify is using a WaitGroup attached to metadata_updated callback
to check if a track is loaded.

Sometimes metadata_updated gets called twice in a very short space of
time and there is no enough time to WaitGroup.Done and stop listen for
metadata_updated between the calls. When that happens WaitGroup.Done is
called twice and a panic is thrown because we're expecting only one
call.

The solution is to use a Once to not call WaitGroup.Done twice.